Use spawn instead of exec.
exec() requires you to quote commands and arguments properly which is
hard to do in a cross-platform way correctly. spawn() requires more
work to handle output but avoids the need to quote commands correctly.

On Windows, commands (including node.exe) generally are accessible via
paths with spaces in them. This helps with #109.