fix #268: Revert "fix #246: remove any double quotes or single quotes from os.tmpdir also sanitize dir option, the template option and the name option"
This reverts commit c8823e549280e11697a510184a69b63bf5bfef7a.

Single quotes must not be removed from paths because they are valid (even if hard to use)
on all OSes. Double quotes are only disallowed on Windows, but tmp should not change any
arg it gets; instead, it should rely on the underlying fs API to fail with an error
that the user needs to fix.